requirements.txt
# FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic>=2.9.0
pydantic-settings>=2.1.0

# FashionCLIP and ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=10.0.0
numpy>=1.24.0

# Vector Search
faiss-cpu>=1.7.4
# Use faiss-gpu if you have CUDA:
# faiss-gpu>=1.7.4

# Database
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
supabase>=2.0.0

# Image Processing
python-multipart>=0.0.6
aiofiles>=23.2.1

# HTTP Client
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring
prometheus-client>=0.19.0

